requirements.txt
pyyaml
wand
cssmin
slimit
ply==3.4